fix bugs && rewrite code
