+++ server/src/app.module.ts
@@ -16,7 +16,8 @@ import { GroupModule } from './modules/group/group.module'
 // -- добавить сериализацию к возвращаемым данным
 // -- вынести декораторы сваггера
 // -- вынести swagger response по модулям
-// -- вынести messages из core (оставить только основные)
+// -- добавить 204 при удалении
+// -- потестить 204 на логин
 
 @Module({
 	imports: [
