+++ server/src/core/swagger/index.ts
@@ -1,2 +1,1 @@
 export * from './messages.enum'
-export * from './responses'
