+++ server/src/modules/staff/swagger/index.ts
@@ -1,1 +1,2 @@
 export * from './messages.enum'
+export * from './responses'
